requirements.txt
nltk
termcolor
pyyaml

